tests/wpt/mozilla/tests/mozilla/htmllabel-activation.html

<!doctype html>
<meta charset="utf-8">
<title>Click action on label</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="test"></div>
<iframe src=htmllabel-form-activation.html id="foo" name="foo"></iframe>

<script>

    var div = document.getElementById("test");
    test(function() {
        document.querySelector("#test").innerHTML = '';
        var label = document.createElement("label");
        var checkbox1 = document.createElement("input");
        checkbox1.type = "checkbox";
        var checkbox2 = document.createElement("input");
        checkbox2.type = "checkbox";

        div.appendChild(label);
        label.appendChild(checkbox1);
        label.appendChild(checkbox2);
        
        label.click();
        assert_true(checkbox1.checked, "checkbox should be checked")
        assert_false(checkbox2.checked, "checkbox should not be checked")

        label.click();
        assert_false(checkbox1.checked, "checkbox should not be checked")
        assert_false(checkbox2.checked, "checkbox should not be checked")
    }, "Check if click on label selects 1st item (checkbox button)");


    test(function() {
        document.querySelector("#test").innerHTML = '<label id="label">hi \
            <input type="checkbox" disabled id="disabledcb"> \
            <input type="checkbox" id="checkbox"></label>';

        var label = document.getElementById('label'),
            disabledcb = document.getElementById('disabledcb'),
            checkbox = document.getElementById('checkbox');
        
        label.click();
        assert_false(disabledcb.checked, "checkbox should not be checked")
        assert_false(checkbox.checked, "checkbox should not be checked")
    }, "If label's 1st child is disabled, click should have no impact");


    // IGNORE HERE DOWN FOR NOW 

    
    var label = document.createElement("label");

    //create a variety of element types to add to the label
    var checkbox1 = document.createElement("input");
    checkbox1.type = "checkbox";
    var checkbox2 = document.createElement("input");
    checkbox2.type = "checkbox";
    checkbox2.id = "checkbox2";
    var disabledCheckbox = document.createElement("input");
    disabledCheckbox.type = "checkbox";
    disabledCheckbox.disabled = true;
    var anchor = document.createElement("a");
    anchor.href = "about:blank";
    var radio = document.createElement("input");
    radio.type = "radio";
    var disabledRadio = document.createElement("input");
    disabledRadio.type = "radio";
    disabledRadio.disabled = true;
    var submit = document.createElement("input");
    submit.type = "submit";
    var submitClicked = false;
    submit.onclick = function() { submitClicked = true  };
    var disabledSubmit = document.createElement("input");
    disabledSubmit.type = "submit";
    disabledSubmit.disabled = true;
    var disabledSubmitClicked = false;
    disabledSubmit.onclick = function() { disabledSubmitClicked = true };

    function resetAll() {
        document.querySelector("#test").innerHTML = '';
        label.removeAttribute("htmlFor");
        while (label.childElementCount != 0) {
            label.removeChild(label.firstChild)
        }
        checkbox1.checked = false;
        checkbox2.checked = false;
        disabledCheckbox.checked = false;
        radio.checked = false;
        disabledRadio.checked = false;
        submitClicked = false;
        disabledSubmitClicked = false;
    }

    test(function() {
        resetAll()
        div.appendChild(label);
        label.appendChild(anchor);
        label.appendChild(checkbox1);

        label.click();
        assert_true(checkbox1.checked, "checkbox should be checked")

        label.click();
        assert_false(checkbox1.checked, "checkbox should not be checked")
    }, "Label's 1st labelable element should be clicked");

    test(function() {
        resetAll()
        div.appendChild(label)
        label.appendChild(radio)
        label.appendChild(checkbox1)

        label.click()
        assert_true(radio.checked, "radio should be checked")
        assert_false(checkbox1.checked, "checkbox should not be checked")
    }, "Check if click on label selects 1st item (radio button)");

    test(function() {
        resetAll()
        div.appendChild(label);
        label.appendChild(disabledRadio);
        label.appendChild(checkbox1);

        label.click();
        assert_false(disabledRadio.checked, "disabled radio should not be checked")
        assert_false(checkbox1.checked, "checkbox should not be checked")
    }, "If label's 1st child is disabled, click should have no impact");


    test(function() {
        resetAll()
        div.appendChild(label);
        label.appendChild(submit);
        label.appendChild(checkbox1);
        
        label.click();
        assert_true(submitClicked, "submit should be selected")
        assert_false(checkbox1.checked, "checkbox should not be checked")
    }, "Check if click on label selects 1st item (submit)");

    test(function() {
        resetAll()
        div.appendChild(label);
        label.appendChild(disabledSubmit);
        label.appendChild(checkbox1);
        
        label.click();
        assert_false(disabledSubmitClicked, "disabled submit should not have been selected")
        assert_false(checkbox1.checked, "checkbox should not be checked")
    }, "If label's 1st child is disabled, click should have no impact");

    test(function() {
        resetAll();
        div.appendChild(label);
        label.appendChild(checkbox1);
        label.appendChild(checkbox2);
        label.htmlFor = "checkbox2";

        label.click();
        assert_false(checkbox1.checked, "checkbox should not be checked");
        assert_true(checkbox2.checked, "checkbox should be checked");
        label.click();
        assert_false(checkbox1.checked, "checkbox should not be checked");
        assert_false(checkbox2.checked, "checkbox should not be checked");
    }, "Label 'for' logic should take precedence over nesting logic");

    test(function() {
        resetAll();
        div.appendChild(label);
        div.appendChild(checkbox1);
        div.appendChild(checkbox2);
        label.htmlFor = "checkbox2";

        label.click();
        assert_false(checkbox1.checked, "checkbox should not be checked");
        assert_true(checkbox2.checked, "checkbox should be checked");
        label.click();
        assert_false(checkbox1.checked, "checkbox should not be checked");
        assert_false(checkbox2.checked, "checkbox should not be checked");
    }, "Verify non-nested label 'for' logic");

    async_test(function(t) {
        var loaded = false;
        var iframe = document.getElementById('foo');
        iframe.onload = t.step_func(function() {
            if (!loaded) {
                loaded = true;
                iframe.contentDocument.getElementById('tenth').click();
            } else {
                t.done();
            }
        });
    }, "label click should flow down to form");

</script>
